reafactor: after treating all compilation warnings as errors, change code to solve errors

+++ exCeption.c
@@ -5,13 +5,13 @@
 #include "exCeption.h" // "" so the preprocessor looks first in the directories relative to this project
 
 const char* STR_PART_ZERO = "An exception occurred, trace:\n";
-const int STR_PART_ZERO_SIZE = 30;
+const long unsigned int STR_PART_ZERO_SIZE = 30;
 const char* STR_PART_ONE = "-> In function \'";
-const int STR_PART_ONE_SIZE = 16;
+const long unsigned int STR_PART_ONE_SIZE = 16;
 const char* STR_PART_TWO = "\': ";
-const int STR_PART_TWO_SIZE = 3;
+const long unsigned int STR_PART_TWO_SIZE = 3;
 const char* STR_PART_THREE = ".\n";
-const int STR_PART_THREE_SIZE = 2; // TODO consider moving these global definitions somewhere else
+const long unsigned int STR_PART_THREE_SIZE = 2; // TODO consider moving these global definitions somewhere else
 
 /* # Public functions */
 
@@ -52,7 +52,7 @@ char* exc_to_str(const exc p_exception){
     }
     
     // allocate memory for the strings. Allocated enough memory for the strings associated to this exception and all of its parents
-    int mem_to_allocate = STR_PART_ZERO_SIZE + exc_str_len(p_exception) + 1; // add one for terminator char
+    long unsigned int mem_to_allocate = STR_PART_ZERO_SIZE + exc_str_len(p_exception) + 1; // add one for terminator char
     char* p_char_to_return = (char*)malloc(mem_to_allocate);
 
     if(p_char_to_return == EXC_NULL_POINTER){
@@ -133,9 +133,9 @@ exc_root* exc_create_root(void){
     }
 }
 
-int exc_str_len(const exc_root* p_exception){
+long unsigned int exc_str_len(const exc_root* p_exception){
 
-    int this_exc_str_len = STR_PART_ONE_SIZE + p_exception->fun_name.length + STR_PART_TWO_SIZE + p_exception->description.length + STR_PART_THREE_SIZE;
+    long unsigned int this_exc_str_len = STR_PART_ONE_SIZE + p_exception->fun_name.length + STR_PART_TWO_SIZE + p_exception->description.length + STR_PART_THREE_SIZE;
 
     if (p_exception->p_parent == EXC_NULL_POINTER) {
         // exception has no parent, return just the length of strings of this exception
@@ -150,7 +150,7 @@ void exc_add_own_str_to_str(const exc_root* p_exception, char* p_first_char){
 
     // since strings are all known in size, use string copy instead of string cat. It is more efficient.
 
-    int tempOffset = 0;
+    long unsigned int tempOffset = 0;
     memcpy(p_first_char, STR_PART_ONE, STR_PART_ONE_SIZE);
     tempOffset = STR_PART_ONE_SIZE;
     memcpy(p_first_char + tempOffset, p_exception->fun_name.p_first_char, p_exception->fun_name.length); // add function name string
